Fix eto view people slider (#1295)
