+++ app/components/shared/PeopleSwiperWidget.module.scss
@@ -5,10 +5,13 @@
 
 .next,
 .prev {
-  border: 1px solid currentColor;
   position: absolute;
   top: 50%;
   transform: translateY(-50%);
+
+  svg {
+    display: block;
+  }
 }
 
 .prev {
